pyproject.toml
[tool.poetry]
name = "google-adk-professional-development-agent"
version = "0.1.0"
description = ""
authors = ["kubikb <kubikbalint@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-adk = "1.3.0"
google-cloud-bigquery = "^3.34.0"
google-cloud-discoveryengine = "^0.13.9"
google-cloud-aiplatform = {extras = ["adk", "agent-engines"], version = "^1.97.0"}


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.4.1"
google-cloud-aiplatform = {extras = ["evaluation"], version = "^1.98.0"}
tabulate = "^0.9.0"
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
